+++ bindings/wasm/examples/06_simple_message.js
@@ -1,0 +1,11 @@
+async function run() {
+    const { ClientBuilder } = require('../node')
+
+    // client will connect to testnet by default
+    const client = await new ClientBuilder().build();
+
+    const message = await client.message().submit();
+    console.log(message);
+}
+
+run()