+++ examples/get_funds.rs
@@ -24,7 +24,7 @@ struct FaucetResponse {
 async fn main() {
     // Create a client instance
     let iota = Client::builder()
-        .with_node("https://api.hornet-1.testnet.chrysalis2.com") // Insert the node here
+        .with_node("https://api.thin-hornet-0.h.chrysalis-devnet.iota.cafe") // Insert the node here
         .unwrap()
         .finish()
         .await