+++ examples/indexation.rs
@@ -9,7 +9,7 @@ use iota_client::{bee_message::prelude::Payload, Client};
 async fn main() {
     // Create a client instance
     let iota = Client::builder()
-        .with_node("https://api.lb-0.testnet.chrysalis2.com") // Insert your node URL here
+        .with_node("https://api.lb-0.h.chrysalis-devnet.iota.cafe") // Insert your node URL here
         .unwrap()
         .with_node_sync_disabled()
         .finish()